+++ magpylib/_src/display/backend_matplotlib.py
@@ -249,6 +249,7 @@ def display_matplotlib(
     """Display objects and paths graphically using the matplotlib library."""
     frames = data["frames"]
     ranges = data["ranges"]
+    labels = data["labels"]
 
     fig_kwargs = {} if not fig_kwargs else fig_kwargs
     fig_kwargs = {"dpi": 80, **fig_kwargs}
@@ -352,10 +353,11 @@ def display_matplotlib(
         for row_col_num, ax in axes.items():
             count = count_with_labels.get(row_col_num, 0)
             if ax.name == "3d":
-                ax.set(
-                    **{f"{k}label": f"{k} (m)" for k in "xyz"},
-                    **{f"{k}lim": r for k, r in zip("xyz", ranges)},
-                )
+                if row_col_num in ranges:
+                    ax.set(
+                        **{f"{k}label": labels[row_col_num][k] for k in "xyz"},
+                        **{f"{k}lim": r for k, r in zip("xyz", ranges[row_col_num])},
+                    )
                 ax.set_box_aspect(aspect=(1, 1, 1))
                 if 0 < count <= legend_maxitems:
                     lg_kw = {"bbox_to_anchor": (1.04, 1), "loc": "upper left"}