+++ magpylib/_src/display/backend_plotly.py
@@ -75,15 +75,17 @@ def match_args(ttype: str):
     return set(named_args)
 
 
-def apply_fig_ranges(fig, ranges, apply2d=True):
+def apply_fig_ranges(fig, ranges_rc, labels_rc, apply2d=True):
     """This is a helper function which applies the ranges properties of the provided `fig` object
-    according to a provided ranges. All three space direction will be equal and match the
-    maximum of the ranges needed to display all objects, including their paths.
+    according to a provided ranges for each subplot. All three space direction will be equal and
+    match the maximum of the ranges needed to display all objects, including their paths.
 
     Parameters
     ----------
-    ranges: array of dimension=(3,2)
+    ranges_rc: dict of arrays of dimension=(3,2)
         min and max graph range
+    labels_rc: dict of dicts
+        contains a dict with 'x', 'y', 'z' keys and respective labels as strings for each subplot
 
     apply2d: bool, default = True
         applies fixed range also on 2d traces
@@ -92,15 +94,27 @@ def apply_fig_ranges(fig, ranges, apply2d=True):
     -------
     None: NoneType
     """
-    fig.update_scenes(
-        **{
-            f"{k}axis": {"range": ranges[i], "autorange": False, "title": f"{k} (m)"}
-            for i, k in enumerate("xyz")
-        },
-        aspectratio={k: 1 for k in "xyz"},
-        aspectmode="manual",
-        camera_eye={"x": 1, "y": -1.5, "z": 1.4},
-    )
+    for rc, ranges in ranges_rc.items():
+        row, col = rc
+        labels = labels_rc.get(rc, {k: "" for k in "xyz"})
+        kwargs = {
+            **{
+                f"{k}axis": {
+                    "range": ranges[i],
+                    "autorange": False,
+                    "title": labels[k],
+                }
+                for i, k in enumerate("xyz")
+            },
+            "aspectratio": {k: 1 for k in "xyz"},
+            "aspectmode": "manual",
+            "camera_eye": {"x": 1, "y": -1.5, "z": 1.4},
+        }
+
+        # pylint: disable=protected-access
+        if fig._grid_ref is not None:
+            kwargs.update({"row": row, "col": col})
+        fig.update_scenes(**kwargs)
     if apply2d:
         apply_2d_ranges(fig)
 
@@ -274,7 +288,6 @@ def process_extra_trace(model):
 
 def display_plotly(
     data,
-    zoom=1,
     canvas=None,
     renderer=None,
     return_fig=False,
@@ -345,11 +358,13 @@ def display_plotly(
                 rows=rows_list,
                 cols=cols_list,
             )
-        ranges = data["ranges"]
+        ranges_rc = data["ranges"]
         if extra_data:
-            ranges = get_scene_ranges(*frames[0]["data"], zoom=zoom)
+            ranges_rc = get_scene_ranges(*frames[0]["data"])
         if update_layout:
-            apply_fig_ranges(fig, ranges, apply2d=isanimation)
+            apply_fig_ranges(
+                fig, ranges_rc, labels_rc=data["labels"], apply2d=isanimation
+            )
             fig.update_layout(
                 legend_itemsizing="constant",
                 # legend_groupclick="toggleitem",