+++ magpylib/_src/display/display.py
@@ -11,14 +11,15 @@ from magpylib._src.defaults.defaults_utility import _DefaultValue
 from magpylib._src.defaults.defaults_utility import get_defaults_dict
 from magpylib._src.display.traces_generic import MagpyMarkers
 from magpylib._src.display.traces_generic import get_frames
+from magpylib._src.display.traces_utility import DEFAULT_ROW_COL_PARAMS
+from magpylib._src.display.traces_utility import linearize_dict
 from magpylib._src.display.traces_utility import process_show_input_objs
 from magpylib._src.input_checks import check_format_input_backend
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.input_checks import check_input_animation
-from magpylib._src.input_checks import check_input_zoom
 from magpylib._src.utility import check_path_format
 
-disp_args = get_defaults_dict("display").keys()
+disp_args = set(get_defaults_dict("display"))
 
 
 class RegisteredBackend:
@@ -30,14 +31,14 @@ class RegisteredBackend:
         self,
         *,
         name,
-        show_func_getter,
+        show_func,
         supports_animation,
         supports_subplots,
         supports_colorgradient,
         supports_animation_output,
     ):
         self.name = name
-        self.show_func_getter = show_func_getter
+        self.show_func = show_func
         self.supports = {
             "animation": supports_animation,
             "subplots": supports_subplots,
@@ -54,7 +55,6 @@ class RegisteredBackend:
         cls,
         *objs,
         backend,
-        zoom=0,
         title=None,
         max_rows=None,
         max_cols=None,
@@ -83,12 +83,18 @@ class RegisteredBackend:
                     f"\nFalling back to: {params}"
                 )
                 kwargs.update(params)
-        frame_kwargs = {
+        display_kwargs = {
             k: v
             for k, v in kwargs.items()
-            if any(k.startswith(arg) for arg in disp_args)
+            if any(k.startswith(arg) for arg in disp_args - {"style"})
         }
-        kwargs = {k: v for k, v in kwargs.items() if k not in frame_kwargs}
+        style_kwargs = {k: v for k, v in kwargs.items() if k.startswith("style")}
+        style_kwargs = linearize_dict(style_kwargs, separator="_")
+        kwargs = {
+            k: v
+            for k, v in kwargs.items()
+            if (k not in display_kwargs and k not in style_kwargs)
+        }
         backend_kwargs = {
             k[len(backend) + 1 :]: v
             for k, v in kwargs.items()
@@ -117,13 +123,12 @@ class RegisteredBackend:
             objs,
             supports_colorgradient=self.supports["colorgradient"],
             backend=backend,
-            zoom=zoom,
             title=title,
-            **frame_kwargs,
+            style_kwargs=style_kwargs,
+            **display_kwargs,
         )
-        return self.show_func_getter()(
+        return self.show_func(
             data,
-            zoom=zoom,
             max_rows=max_rows,
             max_cols=max_cols,
             subplot_specs=subplot_specs,
@@ -136,12 +141,9 @@ class RegisteredBackend:
 def get_show_func(backend):
     """Return the backend show function"""
     # defer import to show call. Importerror should only fail if unavalaible backend is called
-    return lambda: getattr(
+    return lambda *args, backend=backend, **kwargs: getattr(
         import_module(f"magpylib._src.display.backend_{backend}"), f"display_{backend}"
-    )
-
-
-ROW_COL_SPECIFIC_NAMES = ("row", "col", "output", "sumup", "pixel_agg", "in_out")
+    )(*args, **kwargs)
 
 
 def infer_backend(canvas):
@@ -182,7 +184,6 @@ def _show(
     *objects,
     backend=None,
     animation=False,
-    zoom=0,
     markers=None,
     **kwargs,
 ):
@@ -193,9 +194,10 @@ def _show(
 
     # process input objs
     objects, obj_list_flat, max_rows, max_cols, subplot_specs = process_show_input_objs(
-        objects, **{k: v for k, v in kwargs.items() if k in ROW_COL_SPECIFIC_NAMES}
+        objects,
+        **{k: v for k, v in kwargs.items() if k in DEFAULT_ROW_COL_PARAMS},
     )
-    kwargs = {k: v for k, v in kwargs.items() if k not in ROW_COL_SPECIFIC_NAMES}
+    kwargs = {k: v for k, v in kwargs.items() if k not in DEFAULT_ROW_COL_PARAMS}
     kwargs["max_rows"], kwargs["max_cols"] = max_rows, max_cols
     kwargs["subplot_specs"] = subplot_specs
 
@@ -204,7 +206,6 @@ def _show(
 
     # input checks
     backend = check_format_input_backend(backend)
-    check_input_zoom(zoom)
     check_input_animation(animation)
     check_format_input_vector(
         markers,
@@ -216,15 +217,7 @@ def _show(
     )
 
     if markers:
-        objects = [
-            *objects,
-            {
-                "objects": [MagpyMarkers(*markers)],
-                "row": 1,
-                "col": 1,
-                "output": "model3d",
-            },
-        ]
+        objects.append({"objects": [MagpyMarkers(*markers)], **DEFAULT_ROW_COL_PARAMS})
 
     if backend == "auto":
         backend = infer_backend(kwargs.get("canvas", None))
@@ -232,7 +225,6 @@ def _show(
     return RegisteredBackend.show(
         backend=backend,
         *objects,
-        zoom=zoom,
         animation=animation,
         **kwargs,
     )
@@ -407,9 +399,9 @@ def show(
         }
     )
     if ctx.isrunning:
-        rco = {k: v for k, v in kwargs.items() if k in ROW_COL_SPECIFIC_NAMES}
+        rco = {k: v for k, v in kwargs.items() if k in DEFAULT_ROW_COL_PARAMS}
         ctx.kwargs.update(
-            {k: v for k, v in kwargs.items() if k not in ROW_COL_SPECIFIC_NAMES}
+            {k: v for k, v in kwargs.items() if k not in DEFAULT_ROW_COL_PARAMS}
         )
         ctx_objects = tuple({**o, **rco} for o in ctx.objects_from_ctx)
         objects, *_ = process_show_input_objs(ctx_objects + objects, **rco)
@@ -452,11 +444,11 @@ def show_context(
     )
     try:
         ctx.isrunning = True
-        rco = {k: v for k, v in kwargs.items() if k in ROW_COL_SPECIFIC_NAMES}
+        rco = {k: v for k, v in kwargs.items() if k in DEFAULT_ROW_COL_PARAMS}
         objects, *_ = process_show_input_objs(objects, **rco)
         ctx.objects_from_ctx += tuple(objects)
         ctx.kwargs.update(
-            {k: v for k, v in kwargs.items() if k not in ROW_COL_SPECIFIC_NAMES}
+            {k: v for k, v in kwargs.items() if k not in DEFAULT_ROW_COL_PARAMS}
         )
         yield ctx
         ctx.show_return_value = _show(*ctx.objects, **ctx.kwargs)
@@ -491,7 +483,7 @@ ctx = DisplayContext()
 
 RegisteredBackend(
     name="matplotlib",
-    show_func_getter=get_show_func("matplotlib"),
+    show_func=get_show_func("matplotlib"),
     supports_animation=True,
     supports_subplots=True,
     supports_colorgradient=False,
@@ -501,7 +493,7 @@ RegisteredBackend(
 
 RegisteredBackend(
     name="plotly",
-    show_func_getter=get_show_func("plotly"),
+    show_func=get_show_func("plotly"),
     supports_animation=True,
     supports_subplots=True,
     supports_colorgradient=True,
@@ -510,7 +502,7 @@ RegisteredBackend(
 
 RegisteredBackend(
     name="pyvista",
-    show_func_getter=get_show_func("pyvista"),
+    show_func=get_show_func("pyvista"),
     supports_animation=True,
     supports_subplots=True,
     supports_colorgradient=True,