+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [Unreleased] - YYYY-MM-DD
+- Added `units_length` input to the `show` function to allow displaying axes with different length units. This parameter can be set individually for each subplot. ([#786](https://github.com/magpylib/magpylib/pull/786))
+
 ## [5.0.4] - 2024-06-18
 - Add support for Numpy 2.0 ([#795](https://github.com/magpylib/magpylib/pull/789))
 - Fix markers legend not being suppressible ([#795](https://github.com/magpylib/magpylib/pull/789))